# frontend/public/openapi.json
{
  "components": {
    "schemas": {
      "AgentIn": {
        "description": "Agent(design contract,取代 003 的 Profile 角色):人设 + rubric + engine + 出题策略 + self_bookable。\n\n**不再内嵌 questions** —— 题目外置到 QuestionBank,Agent 经 question_strategy 决定怎么出题。\nrubric 归 Agent(评分是 Agent 的职责定义)。default_question_bank_id 供 staff 自助预约用\n(staff 不选题库,用 Agent 预设;admin/HR 发起可覆盖)。",
        "properties": {
          "avatar_style": {
            "anyOf": [
              {
                "enum": [
                  "minimal",
                  "round",
                  "tech",
                  "waveform"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Style"
          },
          "default_question_bank_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Question Bank Id"
          },
          "engine": {
            "$ref": "#/components/schemas/EngineParams"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "question_strategy": {
            "default": "sequential",
            "enum": [
              "sequential",
              "random_n",
              "easy_to_hard",
              "random_n_easy_to_hard"
            ],
            "title": "Question Strategy",
            "type": "string"
          },
          "rubric": {
            "$ref": "#/components/schemas/Rubric"
          },
          "self_bookable": {
            "default": false,
            "title": "Self Bookable",
            "type": "boolean"
          },
          "show_subtitles": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Subtitles"
          },
          "speaker_lock": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Lock"
          },
          "strategy_n": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy N"
          },
          "system_prompt": {
            "default": "",
            "title": "System Prompt",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "AgentIn",
        "type": "object"
      },
      "AgentOut": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "avatar_style": {
            "anyOf": [
              {
                "enum": [
                  "minimal",
                  "round",
                  "tech",
                  "waveform"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Style"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by_admin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Admin"
          },
          "created_by_client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Client"
          },
          "default_question_bank_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Question Bank Id"
          },
          "engine": {
            "$ref": "#/components/schemas/EngineParams"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "question_strategy": {
            "default": "sequential",
            "enum": [
              "sequential",
              "random_n",
              "easy_to_hard",
              "random_n_easy_to_hard"
            ],
            "title": "Question Strategy",
            "type": "string"
          },
          "rubric": {
            "$ref": "#/components/schemas/Rubric"
          },
          "self_bookable": {
            "default": false,
            "title": "Self Bookable",
            "type": "boolean"
          },
          "show_subtitles": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Subtitles"
          },
          "speaker_lock": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Lock"
          },
          "status": {
            "default": "active",
            "title": "Status",
            "type": "string"
          },
          "strategy_n": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy N"
          },
          "system_prompt": {
            "default": "",
            "title": "System Prompt",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "updated_by_admin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By Admin"
          },
          "updated_by_client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By Client"
          },
          "version": {
            "default": "v1",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "name",
          "agent_id"
        ],
        "title": "AgentOut",
        "type": "object"
      },
...
      "AgentVersionsOut": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "current_version": {
            "title": "Current Version",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/AgentOut"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "agent_id",
          "current_version",
          "versions"
        ],
        "title": "AgentVersionsOut",
        "type": "object"
      },
...
      "EngineParams": {
        "properties": {
          "engine_type": {
            "const": "three_stage",
            "default": "three_stage",
            "title": "Engine Type",
            "type": "string"
          },
          "language": {
            "default": "zh-CN",
            "title": "Language",
            "type": "string"
          },
          "llm_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model Id"
          },
          "max_duration_s": {
            "default": 1800,
            "title": "Max Duration S",
            "type": "integer"
          },
          "max_turns": {
            "default": 9999,
            "title": "Max Turns",
            "type": "integer"
          },
          "tts_provider": {
            "anyOf": [
              {
                "enum": [
                  "gpu_omnivoice",
                  "minimax"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tts Provider"
          },
          "voice": {
            "anyOf": [
              {
                "enum": [
                  "male_std",
                  "female_std"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          }
        },
        "title": "EngineParams",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Rubric": {
        "description": "rubric 两形态统一建模 —— 二选一(design contract)。\n\n- per_question_check:逐题 ✓/✗ + 通过线(pass_threshold ∈ [0,1]),典型培训 check。\n- dimension_score:维度打分(dimensions[]),典型面试。\n校验在 model_validator:形态合法 + 对应字段齐备 + 权重/通过线合法。",
        "properties": {
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/RubricDimension"
            },
            "title": "Dimensions",
            "type": "array"
          },
          "mode": {
            "default": "per_question_check",
            "enum": [
              "per_question_check",
              "dimension_score"
            ],
            "title": "Mode",
            "type": "string"
          },
          "pass_threshold": {
            "default": 0.8,
            "title": "Pass Threshold",
            "type": "number"
          }
        },
        "title": "Rubric",
        "type": "object"
      },
      "RubricDimension": {
        "description": "dimension_score 形态的单个维度(典型面试)。",
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "max_score": {
            "default": 5.0,
            "title": "Max Score",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "weight": {
            "default": 1.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "name"
        ],
        "title": "RubricDimension",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "VivaVoce control plane (real-time voice oral-exam). All /api/* endpoints require authentication.",
    "title": "VivaVoce Orchestrator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/agents": {
      "get": {
        "operationId": "list_agents_api_agents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentOut"
                  },
                  "title": "Response List Agents Api Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Agents",
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "create_agent_api_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/{agent_id}": {
      "delete": {
        "description": "删除 Agent(design contract,仅 admin)。\n\n不存在 → 404;仍被**活动会话**(scheduled/in_progress)或**招聘时段 Slot**引用 → 409\n(先处理引用再删,避免悬挂引用)。终态/历史会话不挡(已快照版本)。",
        "operationId": "delete_agent_api_agents__agent_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Agent",
        "tags": [
          "agents"
        ]
      },
      "get": {
        "operationId": "get_agent_api_agents__agent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Agent",
        "tags": [
          "agents"
        ]
      },
      "put": {
        "description": "改版不覆盖历史(design contract):当前版本入 version_history、version bump。",
        "operationId": "update_agent_api_agents__agent_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/{agent_id}/versions": {
      "get": {
        "description": "版本历史(design contract):当前版本 + 历史快照,供报告按发起时版本回溯依据。",
        "operationId": "get_agent_versions_api_agents__agent_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentVersionsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Agent Versions",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/integration/agents": {
      "get": {
        "description": "机器列出全部 Agent(admin 级,非 self_bookable_only —— 与 admin Web 后台一致)。",
        "operationId": "api_list_agents_api_integration_agents_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentOut"
                  },
                  "title": "Response Api List Agents Api Integration Agents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api List Agents",
        "tags": [
          "integration-api"
        ]
      },
      "post": {
        "description": "机器建 Agent(复用 admin 路由逻辑:default_question_bank_id 存在性校验 + 打审计字段)。",
        "operationId": "api_create_agent_api_integration_agents_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Create Agent",
        "tags": [
          "integration-api"
        ]
      }
    },
    "/api/integration/agents/{agent_id}": {
      "delete": {
        "description": "机器删 Agent(复用 assert_agent_deletable:活动会话 / 时段 Slot 引用 → 409)。",
        "operationId": "api_delete_agent_api_integration_agents__agent_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Delete Agent",
        "tags": [
          "integration-api"
        ]
      },
      "get": {
        "operationId": "api_get_agent_api_integration_agents__agent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Get Agent",
        "tags": [
          "integration-api"
        ]
      },
      "put": {
        "description": "机器改 Agent(复用 db.update_agent 的版本快照:当前版入 history、bump version)。",
        "operationId": "api_update_agent_api_integration_agents__agent_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Update Agent",
        "tags": [
          "integration-api"
        ]
      }
    },
...
  }
}